GitissueRet/requirements.txt

# Example
langchain>=0.0.340
langchain-core>=0.1.0
langgraph>=0.0.13
langchain-community>=0.0.7
langchain-ollama>=0.0.3
pydantic>=2.0.0
tiktoken>=0.5.2
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
chromadb>=0.4.22
PyGithub==2.6.1